reafactor: 単純化するために一度一枚画像に戻す
