+++ index.html
@@ -9,7 +9,6 @@
     <div>
         画像を選択<br>
         <input type="file" id="inputImages" multiple><br>
-        <canvas id="leftImage" width="414" height="125"></canvas>
         <canvas id="rightImage" width="414" height="125"></canvas><br>
     </div>
     <div>
