+++ src/three/images/camera_image.ts
@@ -1,5 +1,4 @@
 import * as THREE from "three";
-import { createCamera } from "./create_camera";
 
 const renderer = new THREE.WebGLRenderer({
     canvas: document.querySelector('#rightImage') as HTMLCanvasElement
@@ -13,14 +12,25 @@ const camera = createCamera(canvas);
 const light = new THREE.AmbientLight(0xffffff);
 scene.add(light);
 
-export function getRightCanvasInstance() {
+export function getImageCanvasInstance() {
     return canvas;
 }
 
-export function getRightSceneInstance() {
+export function getImageSceneInstance() {
     return scene;
 }
 
-export function tickRightImages() {
+export function tickCameraImage() {
     renderer.render(scene, camera);
+}
+
+function createCamera(canvas: HTMLCanvasElement) {
+    var nearPlane = 0.01;
+    var farPlane = 1000;
+    const fov = 53.83746828060639;
+    const fovRad = (fov / 2) * (Math.PI / 180);
+    const dist = canvas.height / 2 / Math.tan(fovRad);
+    const camera = new THREE.PerspectiveCamera(fov, canvas.width / canvas.height, nearPlane, farPlane);
+    camera.position.z = dist;
+    return camera
 }
